enh: split diann io into separate script; minor PR comments
